backend/requirements.txt
# -------------------------------
# 🚀 API LAYER (api/, main.py)
# -------------------------------
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
starlette==0.36.3
pytorch==2.7.1

# -------------------------------
# ⚙️ CONFIG / PIPELINE (core/)
# -------------------------------
pydantic==2.6.4
python-dotenv==1.0.1

# -------------------------------
# 🎧 AUDIO PROCESSING (features/, utils/)
# -------------------------------
numpy==1.26.4
scipy==1.12.0
librosa==0.10.1
soundfile==0.12.1
pyloudnorm==0.2.0
# -------------------------------
# 🧠 MUSIC INTELLIGENCE (harmony/, rhythm/, structure/)
# -------------------------------
music21==9.1.0

# Optional (advanced audio analysis)
# Comment out if installation fails
# essentia==2.1b6.dev1110

# -------------------------------
# 📊 ANALYSIS ENGINES (analysis/)
# -------------------------------
scikit-learn==1.4.1.post1
joblib==1.3.2

# -------------------------------
# 📈 DATA PROCESSING
# -------------------------------
pandas==2.2.1

# -------------------------------
# 🤖 AI AGENTS (ai_agents/)
# -------------------------------
openai==1.14.3

# -------------------------------
# ⚡ PERFORMANCE (optional but recommended)
# -------------------------------
numba==0.59.0
tqdm==4.66.2
loguru==0.7.2